Fix the bug from previous commit. choosing subFilter, hiding it, collapsing and expanding is now working properly.
changed the appearance of the buttons to have more round corners.
Problem with 'All Categories': when hiding a chosen subFilter and hitting 'All Categories', the subFilter won't be removed from the layout but mSelectedFilters will be cleared, so there is a mess on the screen.
possible solution: unchoosing a main filter will remove all of its subFilters from mSelectedFilters. Then you can't hide a chosen subFilter without removing it.
<b>end</b>