Add zoom feature in PagerDialogFragment with [TouchImageView](https://github.com/MikeOrtiz/TouchImageView).
Delete the generated image (from html) when exiting RecipeActivity.
Change generated image type to png, that in my case have smaller size than jpg because of the large screen ratio percentage of white background.

Fix problem with RecyclerWithFiltersAbstractFragment when detaching and then attaching it to MainActivity. Search and categories querying seem not to connect with the recyclerView.

Update third party libraries.